backend/pyproject.toml
[project]
name = "clipboard-as-chat-backend"
version = "0.1.0"
description = "Clipboard-as-chat backend"
authors = [
    {name = "mrdsx"}
]
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
